Build system: Make Python 3 the default choice
It makes general sense to me that Python developers far prefer to work
on Python 3 rather than the legacy Python 2 so any Python-2 only bugs
left in the latter are likely not going to be addressed.  That general
sense was confirmed in a recent discussion I had with a Python
developer concerning the intermittent corruption of the
python-generated file, bindings/python/Plframe.pyc _in the source
tree_ that I was experiencing with Python 2.  His strong advice was to
attempt to replicate the problem with Python 3 since that is very
likely the only version the Python developers would be willing to find
a fix for.

Also, Python 3 now generally works for us.  Accordingly I am making it
the default choice when both Python 3 and Python 2 are available on
the system with a fallback to Python 2 if that is the only version
available on the system.  but I am allowing users to override that
behaviour if they specify -DFORCE_PYTHON2=ON (which does what its name
implies which is to only look for a Python 2 version).

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie with both Python 2 and 3 installed) by running CMake
with and without -DFORCE_PYTHON2=ON and observing the correct results
on what Python version (and consistent Python library and numpy with
that version) that were found.