Comprehensive test script: Fix additional whitespace bugs and allow the --prefix option to be relative
The bug fixes were to put quotes around all references to the $PATH
variable (since that environment variable can have whitespace in it
which is rare on Linux but fairly common on Windows) and also put
quotes around all use of the pwd command output.

The --prefix option change was to allow the user to specify either an
absolute prefix or a relative (to the directory where the script
invocation occurs) prefix.  The trick is in all cases to convert the
prefix variable (just as soon as it is completely determined by the
script) to absolute form so the rest of the script can continue to use
this prefix variable in an absolute form as previously.

In addition, I quoted all directory locations that are echoed;
collected additional Java, Ada, D, and PWD environment variables;
systematically processed all environment variable collection using
printenv output (to assure it was an environment variable being set
rather than an ordinary variable); summarized which environment
variables were collected in the environment variable log file,
processed the further input to that file with sed so that everything
to the right of the "=" sign is quoted (useful for display of all
path-related information and compiler flags); and fixed a minor issue
(a double space rather than single space to separate components of the
command) with use of the dashboard_option variable.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) using a constrained comprehensive test.  Details of
those constraints to follow in a subsequent commit message for the
rest of the file changes that were tested.