Build system: Fix "space in full pathname" bug for the check_plplot_h.inc target
This target checks the consistency of the OCaml source code files.
For the previous release 5.12.0, I thought I had discovered a bug in
this target so I dropped the dependency of the check_all target on
this target.  However, that "no space" bug was a false alarm which
could be solved by installing the libpcre-ocaml-dev Debian package.
Therefore, in this commit I reinstated that dependency.  However, that
still left a bug for the case where either/both the source tree or
build tree prefix contains a space.  This commit fixes that issue by
copying touchup.ml (the file that is executed by the ocaml
application) from the source to build tree so a relative pathname can
be used in the build tree, and by using a relative pathname for the
result file.  However, it turned out that the plplot_h input file for
the touchup.ml Ocaml application could be accessed with full source
tree path in quoted and escaped form so that is what I did.  But why
these inconsistencies in what is allowed for "spaced" pathnames or
not?

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by configuring PLplot from a source tree with
spaces in its prefix using a build tree with
spaces in its prefix as follows:

cmake "-DCMAKE_INSTALL_PREFIX=/home/software/plplot/installcmake blank" -DBUILD_TEST=ON -DDEFAULT_NO_BINDINGS=ON -DENABLE_ocaml=ON -DDEFAULT_NO_DEVICES=ON -DPLD_ps=ON -DGENERATE_PLPLOT_H_INC=ON "../plplot blank .git" >& cmake.out

(the GENERATE_PLPLOT_H_INC option allows the check_plplot_h.inc target to be built.)
That was followed by a successful build of the check_plplot_h.inc target

This target checks the consistency of the OCaml source code files.
For the previous release 5.12.0, I thought I had discovered a bug in
this target so I dropped the dependency of the check_all target on
this target.  However, that "no space" bug was a false alarm which
could be solved by installing the libpcre-ocaml-dev Debian package.
Therefore, in this commit I reinstated that dependency.  However, that
still left a bug for the case where either/both the source tree or
build tree prefix contains a space.  This commit fixes that issue by
copying touchup.ml (the file that is executed by the ocaml
application) from the source to build tree so a relative pathname can
be used in the build tree, and by using a relative pathname for the
result file.  However, it turned out that the plplot_h input file for
the touchup.ml Ocaml application could be accessed with full source
tree path in quoted and escaped form so that is what I did.  But why
these inconsistencies in what is allowed for "spaced" pathnames or
not?

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by configuring PLplot from a source tree with

cmake "-DCMAKE_INSTALL_PREFIX=/home/software/plplot/installcmake blank" -DBUILD_TEST=ON -DDEFAULT_NO_BINDINGS=ON -DENABLE_ocaml=ON -DDEFAULT_NO_DEVICES=ON -DPLD_ps=ON -DGENERATE_PLPLOT_H_INC=ON "../plplot blank .git" >& cmake.out

That configuration of the PLplot build was followed by a successful build of the
check_plplot_h.inc target, i.e.,

make check_plplot_h.inc

Note the prefixes of the source and build trees (and also the install
tree but that is irrelevant for this particular test) contained spaces
for this successful test.