ocaml binding: drop special handling of plhist array argument
The generic rules in touchup.ml appear to work well for plline (which
has similar array arguments as plhist) so I tried using those generic rules for
touching up the plhist arguments, and the result is different, i.e.,

-[mlname(plhist)] void c_plhist ( int n, [in, size_is(n), size_is(n)] double * data, double datmin, double datmax, int nbin, plplot_hist_style opt );
+[mlname(plhist)] void c_plhist ( int n, [in, size_is(n)] double * data, double datmin, double datmax, int nbin, plplot_hist_style opt );

with the extra size_is(n) removed, but I think that is likely a good
thing.  (In fact, I don't understand how the interface to plhist
worked before.)  So this simplifying commit is a small step in the direction
of understanding what is needed to generate the ocaml binding.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by using the DGENERATE_PLPLOT_H_INC=ON option, building
the check_plplot_h.inc target to check that plplot_h.inc is consistent
with touchup.ml, and building
the test_diff_psc target which generated
the (long-standing) PostScript difference result:

ocaml
  Missing examples            :
  Differing graphical output  :  08 16 19 33
  Missing stdout              :
  Differing stdout            :

That is, this change to plhist introduced no differences in example 5
(the only example that uses plhist).