Build System: Implement the PAUSE_CERTAIN_INTERACTIVE_DEVICES option
This experimental option defaults to OFF, but if the user specifies
-DPAUSE_CERTAIN_INTERACTIVE_DEVICES=ON, then for the wingcc and
wxwidgets devices the -np option is not used for test targets, e.g.,
test_c_wingcc and test_c_wxwidgets.  The reason these two devices are
singled out for this experimental treatment is that they have an issue
of only plotting a blank screen when -np is specified.  So until that
issue for these devices is fixed (i.e., they plot the actual plot
rather than a blank screen when -np is specified), this experimental
-DPAUSE_CERTAIN_INTERACTIVE_DEVICES=ON CMake option has been
implemented to drop the -np option for tests of these devices.  The
result is the plots are rendered properly for tests of these devices
at the expense of much additional user interaction (e.g., hitting the
enter key to get through the plot pages for each example that is run).

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by specifying the CMake option
-DPAUSE_CERTAIN_INTERACTIVE_DEVICES=ON and building the
test_c_wxwidgets target (which as expected displayed all plots
properly without the -np option at the expense of much more
interaction to get through the plots) and test_c_xcairo target (which
as expected displayed all plots properly with the -np option so no
interaction was required to get through the plots).