Increase sophistication of script for trailing whitespace removal
The user is now presented with a list of files that have trailing
whitespace, and is given the option of removing that whitespace
for all those files.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by running the script with and without files with
trailing whitespace and for the "with" case optionally removing that whitespace.