Update the license for the PLplot version of the csa and nn libraries
These libraries are substantially modified versions of the 2003
versons of Pavel Sakov's csa and nn libraries that are used to help
provide our plgriddata API for our core plplot library.  The license
update we have done with this commit for the PLplot versions of these
libraries was to adopt the licensing text now used by Pavel Sakov for
his latest versions of the csa and nn libraries (with the add-on note
concerning the non-free triangle library specifically excluded from
the latter licensing text since our changes to nn make the triangle
library irrelevant to our version of nn.)  This latest licensing text
is a simplification of the 3-clause BSD license, and it is the opinion
of Ole Streicher, that this license is free and compliant with the
DFSG (Debian Free Software Guidelines) which should allow him to close
a critical bug for his Debian package of PLplot.

Since those 2003 libraries were copyright Pavel Sakov we needed his permission to do
this licensing update which he kindly gave us with the following e-mail text:

Date: Wed, 12 Jul 2017 14:08:28 +1000
Subject: Re: Request to update the licensing for our stripped down and substantially modified 2003 versions of nn and csa to your modern licenses for nn and csa
----------------------------------------

Hi Alan,

I agree with the changes in the libraries nn and csa used in PLplot as
proposed in your message and quote below the first paragraph from your
message as you requested to confirm my agreement.

"On behalf of the PLplot developers I request your agreement for
updating the licensing of our stripped-down and substantially modified
versions of your 2003 version of the csa and nn libraries to be
respectfully the same as your modern license for csa given at <
https://github.com/sakov/csa-c/blob/master/csa/LICENSE> and your modern
license for nn given at
<https://github.com/sakov/nn-c/blob/master/nn/LICENSE> (with the last two
note lines concerning triangle dropped because we use the free qhull
library rather than non-free triangle software <
http://www.cs.cmu.edu/~quake/triangle.html>)."

Best regards,
Pavel Sakov