Replace out-dated Debian packaging files with documentation of where to find the latest Debian packaging information
As a result of this change, the debian subdirectory is removed.
Therefore, I have dropped the redundant build-system constraint
that the debian subdirectory is not included in our source release
tarballs.  In addition, I have documented where to find the
latest Debian packaging information in
the new file debian_packaging/README.