Adjust for the new "plplot-doc-check" name of Hǎiliàng Wang's API consistency checker for PLplot
That golang programme checks for consistency between our DocBook documentation of our API
in doc/docbook/src/api.xml and the include/plplot.h header that defines our C API.
Hǎiliàng upgraded that programme today to correctly handle callback function
arguments, and he took the opportunity to change the name of the programme from "check"
(which I had symlinked to a unique name because that name was too generic) to
"plplot-doc-check" (which is sufficiently unique that it doesn't require a symlink).

This commit adjusts for that new name, and I also took this opportunity to document
how the new plplot-doc-check should be built in the comments in this CMakeLists.txt file.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by building plplot-doc-check and putting it on the
PATH, configuring the PLplot
build with the cmake option -DADD_SPECIAL_CONSISTENCY_CHECKING=ON
and checking for consistency between DocBook and C versions of
our API by building the check_api_xml_consistency target.  The
result was _almost_ a succes with two inconsistencies found which
are as follows:

c_plshade1 is not documented.
para 0 of func plGetCursor TYPE mismatch doc[ PLGraphicsIn * ] vs. header[ struct PLGraphicsIn * ]

The first corresponds to a real inconsistency which will be fixed when plshade1 is deprecated
(the documentation is ahead of the C code in that regard).  The second is a spurious warning
which I have brought to the attention of Hǎiliàng Wang.