plmap cruft removal
As of PLplot-5.9.10, a new version of plmap was implemented that used
shapefile format (accessed via shapelib) for maps.  In addition other
powerful map API (see the last page of standard example 19) that
depended on shapelib map data was implemented as well.  However, we
still made the old plmap implemention that depended on the
(undocumented) binary format of our *.map files available when the
user specified -DPL_DEPRECATED=ON, and this arrangement continued
through PLplot-5.13.0.  However, it appears our users are generally
satisfied with the new shapefile-based plmap functionality and we no
longer want to maintain or test that old plmap functionality based on
the *.map format.  So for this commit that old plmap functionality
and associated *.map files have been completely removed from our
source tree.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by building the test_diff_psc and install targets
in the build tree for the shared library case.

I did the above test twice (once as per normal and once with a
temporary local variation of cmake/modules/FindShapelib.cmake that
disabled finding of shapelib).  In both cases there were no obvious
configuration or run-time issues, and the PostScript difference
results had no regressions.  And (as expected) when shapelib finding was
disabled, example 19 generated plot results with no map information
other than meridians (from the calls in that example to plmeridians)
and labels, and the core plplot library (as expected) emitted "PLPLOT
WARNING" messages with the following messages:

plmap is a no-op because shapelib is not available.
plmapfill is a no-op because shapelib is not available.
plmapline is a no-op because shapelib is not available.
plmaptex is a no-op because shapelib is not available.