Remove the rest of the sys subdirectory tree
The "sys" subdirectory tree was originally used in the pre-CMake days
to implement various build systems for various platforms.  Most of
this subdirectory tree was removed in commit 69abf21 (2009-09-23)
followed by the removal of sys/dos/djgpp in commit aaf7332
(2010-01-25).  This left just sys/win-tk.  However, that directory has
not been developed for a decade or more, and my scan through the files
there showed there was nothing left of value there so I have removed
this last of the "sys" subdirectories in this commit.