Massively simplified the map code to fix a rendering bug and added PLFLT_HUGE_VAL to plplot.h
A rendering bug was reported - see email to the list to the plplot-devel
list entitled Plmap omits lines at the left hand side of the map dated
18th Nov 2016. The bug was as described, lines on a lat/lon map that
touched the left edge of the plot dissapeared.

I have removed the code that tried to be clever and wrap cut sections
of lines that weren't supposed to be plotted and attempted to wrap them
round the globe if needed. Instead I now draw objects multiple times if
needed to cover the wraparound. Note we also now support doing a plot
which loops round the globe multiple times, so for example you could do
lon -180 to +540 and you would get two full renderings of the globe.

I added PLFLT_HUGE_VAL as I wanted to ensure that if someone threw one
of the floating point limiting values in as the limit for the map plot
then I wouldn't get stuck in a near infinite loop.