Fixed PLFLT issue in plmap and issue with polygons wrapping the globe
There was a problem with array types if PLFLT was not a double. this got
introduced because shapelib always deals in doubles and in the last
commit the conversion got removed. It has been added back in a more
useful place.

Also dealt with plmap rendering of polygons that wrap the whole globe,
e.g. Antarctica. We should also effectively deal with the case where
such a polgon gets turned inside out, like when doing a polor plot with
the north pole in the centre.