Fixed wxWidgets even odd winding bug when using viewer
This was really a buffer bug or at least the modifications here fix a
buffer bug.
Basically the dev_eofill variable was not being recorded in the buffer
anywhere so it was never passed to the viewer. It is now recorded as
part of the the state at beginning of page and plP_state is now called
if opt_eofill is called after initialization which causes the state
change to be recorded. A new state #define has been added to facilitate
this.

Tested by Phil Rosenberg by running example 27 with and without the
-eofill command argument, however, only up to the point of checking in a
debugger that wxPLViewer receives the value and calls the fill function
with the correct fill option. I'm not actually sure what the difference
in output should be - both looked the same to me, but maybe I'm missing
something subtle.