Fixed plreplot and wxPLViewer bugs regarding begining of pages
This addressed the bug report at
https://sourceforge.net/p/plplot/bugs/168/

There were two issues to address. The first was that if plreplot is
called but the PLStream happens to be at the beginning of the page then
the new beginning of page (which generally involves clearing the
background to the bgcolour set from the buffer) was ignored. Now the
status is set to end of page before the buffer read begins.

The second issue was that during initialisation wxPLViewer added a
beginning of page event to the buffer. It then received the buffer to
plot, which included a second beginning of page even and appended this
to the existing buffer. This gave two consecutive beginning of page
entries in the buffer and the second one (which is the one we actually
want) was ignored. This was fixed by checking if the data
received was the first data, and if so then overwrite rather than append
to the buffer.

Tested by running x16 and checking we get correct background colours for
all plots.