Added support for xor rendering and fixed a locate bug in wxWidgets.
Example 20 didn't do the locate part because wxWidgets driver did not
support xor mode. added a flag to say this is supported, but actually
support isn't really there fully. Support is there when a wxDC is passed
into the driver, however when using the wxPLViewer there is an issue -
when xor mode is set the draw commands are no longer saved to the buffer,
so they never get passed to the viewer. Changing this behaviour will be
discussed on the email list.

A locate bug was then shown up by this example. When locate occurred on
a page other than the first, the PLViewer may have still been on an
earlier page. We now record the page on which the locate occurred and do
not action it until the user has moved to that page.

The three sem comms suffers from a further locate bug which needs
further work and is caused by both the core and the viewer code
getting stuck sitting and waiting for the other to transmit data. But
this commit means that the older comms method deals with locates
absolutely correctly.