Towards correct locate state variable assignment in wxPLViewer
I found that the locate state variable should represent the mouse
buttons that were down immediately before the mouse event in question.
I.e any other buttons held down at the same time or if the event in
question was a release, then the button that was released.

In theory the logic now in wxPLViewer is correct, but the frame is not
capturing button release events so these events are not generating any
locate information. This may require use of a wxPanel which I have been
meanuing to do anyway as it is best practice. This commit is still an
improvement on the incorrect locate information that was being provided
though so I thought it was worth commiting now.