Attempted workaround for wxPLViewer IPC3 hangs
When pllocate is called the core code then requests data from
wxPLViewer in a blocking manner. Meanwhile wxPLViewer received the
locate instruction, but carried on requesting further plot commands from
the core code in a blocking manner while waiting for the user to click
the mouse. This caused deadlock.

Modified the wxPLViewer code so that it no longer requests data from the
core code while waiting for locate data.

This sortof works and sometimes allows execution beyond the use of
locate in example 20, but then we seem to hit deadlock later. Not sure
why.