Set the option of a render delay after resizes on wxPLplotwindows
When rendering complex plot resizes can become jerky, unresposive and in
some cases impossible because the window is continually trying to
rerender the plot.

Now set this up so that a delay to the render can be set and after
resize the render only happens once the size has remained constant for
that delay time. For wxPLviewer that delay time has been set to 1000 ms
(1 s). This fixes bug #170 https://sourceforge.net/p/plplot/bugs/170/.