Fixed recently introduced error that stopped wxPLViewer realizing plotting was complete
A recent change accidentally adjusted the logic for a set of if/else
statements, meaning that the timer that checks for new data got
restarted. The result was a hang with IPC3 comms while the blocking
getData function waited for data that never arrived. For non-IPC3 comms
the repeated checks had no specific bad effects other than wasted cpu
cycles I guess.