Build systems: output additional information
The added information is the following:

* Full pathname of the cmake executable
* Name of generator
* CMAKE_INCLUDE_PATH (both cache variable and environment variable forms)
* CMAKE_LIBRARY_PATH (both cache variable and environment variable forms)
* CMAKE_PLATFORM_INFO_DIR (the location where configured language support files live)

I have consistently added this output both for the core build system
as well as the CMake-based version of the installed examples build system.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by running cmake in the build tree and observing the
output.  The install-tree version has not been tested this way, but
the identical CMake logic has been copied in that case so it should be
fine.