WxWidget bug fix for wxWidgets during initialisation
The wxPLViewer uses the size of the previously rendered bitmap to test
whether or not it needs to rerender on a resize event. However it can
receive a resize event before initialising te bitmap and on some
platforms that causes a crash/assert fail/exception. Added a check for
bitmap initialisation before getting the size of the bitmap.