Build system: fix sanity check logic for set_library_properties function
if(NOT(${library}_SOVERSION OR ${library}_VERSION OR LIB_DIR))

==>

if(NOT(DEFINED ${library}_SOVERSION AND DEFINED ${library}_VERSION AND DEFINED LIB_DIR))

As is obvious above the prior version of the logic that has now been fixed
was completely bogus (i.e,  used OR rather than the correct AND, and used true/false (which
failed when ${library}_SOVERSION was 0) rather than doing the correct check that the
variable had been DEFINED).

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by configuring PLplot with cmake.  The result was
there were no CMake errors due to the above revision in the sanity
checking logic.