C example 01 and Fortran binding and example 01: improvements in command-line parsing
The C example improvement was to implement the -pl_parse_skip option
for examples/c/x01c.c that demonstrated the existing PL_PARSE_SKIP
mode of the C version of plparseopts.  The Fortran binding improvement
was to make similar PL_PARSE_SKIP mode functionality available for
that binding via and added variant of the plparseopts API and a new
Fortran function called plget_arguments.  The
examples/fortran/x01f.f90 improvement was to implement a demonstration
of that capability that gave the same results as the corresponding C
results.  For further details see the write-up concerning these
changes in README.release.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie with gfortran-4.9.2) by building the x01c, x01f, and ps
targets, and running valgrind in the way described in README.release
with perfect results.