PLplot build system: configure library builds in a more uniform manner
I replaced calls to add_library, target_link_libraries, and the
set_library_properties function with a call to the build_library
function which for standard libraries can either be static or shared
(but not STATIC only and not MODULE). This function calls add_library,
creates a namespaced ALIAS library corresponding to that library,
calls target_link_libraries in a standard way, and sets standard
library properties.  Thus, this change significantly reduces
the CMake code used to build our libraries and makes sure such
building is done in a more uniform way.

These new ALIAS libraries will be required in the future (so that the
installed libraries can have namespaced targets associated with them
so that common code between build tree and install tree can use
namespaced library targets).  But the future plan is not to confine
the use of namespaced targets just to the common code but also use
them everywhere in the build tree where a library target is used in a
read-only way as a matter of style (and safety for the
target_link_libraries item case).  So for this commit I used ALIAS
library targets in a small way to test them in
that context.

I also deleted the set_library_properties function since it is no
longer used.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by building the test_noninteractive and
test_interactive targets for shared libraries in the build tree
without any configure, build, run-time, or PostScript difference
issues.