Python binding: Improve UTF-8 fixup
Our swig documentation strings (which include some UTF-8 glyphs such as
"≥") are propagated directly to the swig-generated bindings/python/plplotc.py
Python script.  So years ago when our bindings were strictly Python2 we
had to implement a sed fixup to modify that file to append " It uses coding: utf-8"
to the first line of that file to signal (according to PEP 263) that the encoding
of that file is UTF-8.

This commit improves the implementation of this fixup as follows:

* Changed the implementation from using an additional
  add_custom_command/add_custom_target pair to the much simpler
  add_custom_command(TARGET _plplotc POST_BUILD ...) signature.
  The result is the same; the sed fixup occurs *AFTER*
  whenever swig (re-)generates bindings/python/plplotc.py.

* It turns out that this fixup is not needed for Python3 because
  for that version of Python the default encoding is UTF-8.  Therefore,
  only implement the fixup for the Python2 case.

* Because of the new implementation, the _plplotc_fixed target no
  longer exists.  So replace that target with the  _plplotc target
  wherever it occurs in our build system.

* Change the comments in README.Release_Manager_Cookbook accordingly.

As a result of this commit "_plplotc_fixed"
no longer exists in our source tree except for the historical
ChangeLog.release file.

N.B. Note this remaining small Python2 fixup is actually a workaround
for either a Python2 issue (for defaulting to ascii encoding rather
than UTF-8) or else swig issue for not working around this Python2
issue. But I have decided not to report this issue because of the
uncertainty of where to report it and because it is undoubtedly an
uncommon issue that is getting rarer (because Python2's share of the
Python market keeps dropping).  In fact Python3 is working so well for
PLplot, I plan to deprecate and eventually remove our Python2 support
(including the current fixup) in the intermediate term (i.e., within a
year or two).

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by freshly configuring PLplot with either
the set of CMake options

# The python2 case
-DBUILD_TEST=ON --DDEFAULT_NO_DEVICES=ON -DPLD_ps=ON -DDEFAULT_NO_BINDINGS=ON -DENABLE_python=ON -DFORCE_PYTHON2=ON

or

# The (default) python3 case
-DBUILD_TEST=ON --DDEFAULT_NO_DEVICES=ON -DPLD_ps=ON -DDEFAULT_NO_BINDINGS=ON -DENABLE_python=ON

For both cases, building the test_diff_psc target showed no configuration, build,
run-time, or PostScript difference issues.