Build system: Force DROP_GTK_PLUS_2_BUILDS to be ON
The fundamental purpose of this change is to make sure all our GTK+
dependencies are version 3 rather than version 2 since version 3 has
long been the standard version of GTK+ and a combination of versions 2
and 3 will likely not work in any case.

The practical implication of this change is
examples/c/extXdrawable_demo.c is never built since that app consists
of calls to GTK+ API that was already deprecated in GTK+ version 2 and
completely removed in GTK+ version 3.  In addition, this change means
the PLplot ocaml "cairo" API cannot be used, but that API has already
been disabled due to lack of maintenance.