octave binding and examples; adjust for octave-4 changed interpretation of strings
The octave-4 difference we are adjusting for is that NULL's embedded
in strings no longer terminate output of printf, etc.  Instead, the
entire string is output with embedded NULL's ignored.

To adjust for this change, the octave binding of plplot stores the
results from plgver, plgdev, and plgfnam (all the PLplot API with
output character string arguments) in an ordinary C string, and
inserts the NULL-terminated substring of those results into an octave
charMatrix which is then returned to the octave environment as an
octave string of the correct (NULL-terminated) length.

N.B. Backwards-incompatible change to plgver, plgdev, and plgfnam.

The results for these functions are backwards incompatible because the
result returned to the octave environment is a string rather than the
transpose of a string which was the previous (badly designed)
historical result.  Also note because of bad historical choices there
are many other data returned by the current octave binding for PLplot
where the transpose must be taken and the long-term plan is to
introduce more backwards-incompatible changes by getting rid of all
these unnecessary transposes.  So this commit is the first step toward
that goal.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Buster) by configuring PLplot with cmake 3.11.2 using the
options -DBUILD_TEST=ON -DTRY_OCTAVE4=ON -DUSE_INCRTCL_VERSION_4=ON
and building the test_diff_psc target.  The configuration, build, and
run time results had no obvious errors, and the present commit solved
stdout differences for the octave part of the PostScript difference
report.  Therefore, the exceptions to a perfect PostScript difference
report are now reduced to

octave
  Missing examples            :  26 33
  Differing graphical output  :
  Missing stdout              :
  Differing stdout            :
lua
  Missing examples            :  23
  Differing graphical output  :
  Missing stdout              :
  Differing stdout            :

where those missing examples are a temporary workaround to avoid
exposing upstream bugs in octave 4 (likely 8-bit clean issues with
octave-4 UTF-8 handling of strings) and lua5.3 (see
<https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=902238> for details
of this severe bug).