Build system: Fix lua and octave rebuild bugs
A maintenance quirk for the PLplot build system is that when examples
are (temporarily) dropped (as in previous lua and octave commits to
avoid exposing lua and octave upstream bugs, GRR) it is necessary to
drop those examples from the list_example_files function in
cmake/modules/plplot_functions.cmake to avoid rebuild bugs such as the
ones fixed by this current commit.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Buster) by building the test_diff_psc target twice and
demonstrating for the second of those that there are no unnecessary
rebuilds such as occurred for lua and octave examples prior to this
commit.