Octave binding: For octave-4.4 and above to avoid deprecations warnings replace calls to the is_empty and is_cell methods by calls to the preferred octave-4.4 names for those methods which are isempty and iscell
As of version 4.4, the preferred names for the is_empty and is_cell methods have been changed
to isempty and iscell to make them consistent with the corresponding octave function names.

The current commit uses conditional compilation to use isempty and
iscell for octave-4.4 (to avoid the 4.4 is_empty and is_cell
deprecation warnings), but otherwise it continues to use the old
names.  Therefore, this commit should not introduce issues for
octave-4 versions below 4.4 and likely also octave-3, but I am not in
a position to test those older octave versions.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Buster) by configuring PLplot with the cmake options
-DBUILD_TEST=ON -DTRY_OCTAVE4=ON -DUSE_INCRTCL_VERSION_4=ON and
building the test_noninteractive target with octave 4.4.  All the
previous is_empty and is_cell deprecation build warnings are now gone,
and this test showed no obvious configure, build, or run-time
regressions as a result of this commit.