Octave binding: remove last build warning
The build warning was caused by the following code fragment propagated
to the C++ code generated by swig to implement the octave binding:

// Check if version >= 3.4.0
#if OCTAVE_API_VERSION_NUMBER < 45
    retval( 0 ) = octave_value( local_charMatrix, true );
#else
    retval( 0 ) = octave_value( local_charMatrix );
#endif

This is an attempt to use the first (ancient) octave_value form if the
user has a very old version of octave, but this is inconsistent with
other parts of the code which uses the modern (second) form of
octave_value without attempting to support ancient octave.
Furthermore, even this attempt at backwards incompatibility with
ancient octave is not implemented correctly since
OCTAVE_API_VERSION_NUMBER is not #defined for modern octave so the C++
preprocessor assumes its value is 0, and the first (ancient) form of
the octave_value call is (incorrectly) used with accompanying build
warnings.  I have absolutely no idea how that ancient form still
worked for octave-4.4 prior to this commit, but in any case the
warning issue is solved by this commit by only using the second modern
form of the octave_value call consistent with other code in
plplot_octave.i.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Buster) by configuring PLplot with the cmake options
-DBUILD_TEST=ON -DTRY_OCTAVE4=ON -DUSE_INCRTCL_VERSION_4=ON and
building the test_noninteractive target with octave 4.4.  All
remaining deprecation build warnings for octave are now gone, and this
test showed no obvious configure, build, or run-time regressions as a
result of this commit.